requirements.txt
pillow
onnxruntime
numpy
pooch
tqdm